Remove read and write and combine into data_access - also delete __get_file_format unused functions
