+++ spp/engine/data_access.py
@@ -1,4 +1,4 @@
-
+from spp.utils.query import Query
 
 class DataAccess:
     """
@@ -58,3 +58,75 @@ def set_run_id(df, list_partition_column, run_id, is_spark):
 
         df = df.withColumn('run_id', f.lit(run_id))
     return df
+
+
+
+def spark_write(df, data_target, counter, logger, **kwargs):
+    """
+    Writes a Spark DataFrame to a file.
+    :param counter: Int used to modify file path name if there is already a file
+    existing with the desired name
+    :param data_target: Dictionary containing information on where to save the data
+    :param df: Spark DataFrame
+    :param logger:
+    :param kwargs: Other keyword arguments to pass to df.write.save()
+    """
+    tmp_path = ''
+    if isinstance(counter, int) & (counter >= 1):
+        tmp_path = "/data" + str(counter)
+    data_target['location'] = data_target['location'] + tmp_path
+    write_spark_df_to_s3(df, data_target, logger=logger)
+    logger.debug("Writing to file location: " + data_target['location'])
+
+
+def write_spark_df_to_s3(df, data_target, logger):
+    from pyspark.context import SparkContext
+    from awsglue.context import GlueContext
+    from awsglue.dynamicframe import DynamicFrame
+
+    logger.debug(f"Writing spark dataframe to {repr(data_target)}")
+    glue_context = GlueContext(SparkContext.getOrCreate())
+    dynamic_df_out = DynamicFrame.fromDF(df, glue_context, "dynamic_df_out")
+
+    block_size = 128 * 1024 * 1024
+    page_size = 1024 * 1024
+    glue_context.write_dynamic_frame.from_options(
+        frame=dynamic_df_out,
+        connection_type="s3",
+        connection_options={
+            "path": data_target["location"],
+            "partitionKeys": data_target["partition_by"],
+        },
+        format="glueparquet",
+        format_options={
+            "compression": "snappy",
+            "blockSize": block_size,
+            "pageSize": page_size,
+        },
+    )
+    logger.debug("write complete")
+
+
+def spark_read(spark, cursor, logger, **kwargs):
+    """
+    Reads data into a DataFrame using Spark. If the cursor is an SPP Query,
+    the Spark metastore is used,
+    otherwise the cursor is treated like a file path.
+    :param cursor: Query instance or file path String
+    :param kwargs: Other keyword arguments to pass to spark.read.load()
+    :param logger: SPP logger object to pass to logger functions
+    :param spark: Spark session
+    :returns Spark DataFrame:
+    """
+
+    # If cursor looks like query
+    if isinstance(cursor, Query):
+        logger.debug(f"Reading from database, query: {str(cursor)[:-1]} reader: {spark}")
+        return spark.sql(str(cursor)[:-1])
+
+    # Otherwise, treat as file location
+    else:
+        logger.debug(f"Reading from file {cursor}")
+        return spark.read.load(cursor, format=_get_file_format(cursor), **kwargs)
+
+
