--- spp/engine/write.py
@@ -1,53 +1,0 @@
-
-
-
-def spark_write(df, data_target, counter, logger, **kwargs):
-    """
-    Writes a Spark DataFrame to a file.
-    :param counter: Int used to modify file path name if there is already a file
-    existing with the desired name
-    :param data_target: Dictionary containing information on where to save the data
-    :param df: Spark DataFrame
-    :param logger:
-    :param kwargs: Other keyword arguments to pass to df.write.save()
-    """
-    tmp_path = ''
-    if isinstance(counter, int) & (counter >= 1):
-        tmp_path = "/data" + str(counter)
-    data_target['location'] = data_target['location'] + tmp_path
-    write_spark_df_to_s3(df, data_target, logger=logger)
-    logger.debug("Writing to file location: " + data_target['location'])
-
-
-def write_spark_df_to_s3(df, data_target, logger):
-    from pyspark.context import SparkContext
-    from awsglue.context import GlueContext
-    from awsglue.dynamicframe import DynamicFrame
-
-    logger.debug(f"Writing spark dataframe to {repr(data_target)}")
-    glue_context = GlueContext(SparkContext.getOrCreate())
-    dynamic_df_out = DynamicFrame.fromDF(df, glue_context, "dynamic_df_out")
-
-    block_size = 128 * 1024 * 1024
-    page_size = 1024 * 1024
-    glue_context.write_dynamic_frame.from_options(
-        frame=dynamic_df_out,
-        connection_type="s3",
-        connection_options={
-            "path": data_target["location"],
-            "partitionKeys": data_target["partition_by"],
-        },
-        format="glueparquet",
-        format_options={
-            "compression": "snappy",
-            "blockSize": block_size,
-            "pageSize": page_size,
-        },
-    )
-    logger.debug("write complete")
-
-
-def _get_file_format(location):
-    # ToDo
-    format = "parquet"
-    return format
